fixed query 2 fully

+++ backend/military_unit/src/main/java/ccfit/nsu/ru/daryaevd/military_unit/dto/SoldierDto.java
@@ -20,6 +20,6 @@ public class SoldierDto {
     private Date dateOfIssueMilitaryCard;
     private Long masId;
     private Integer typeOfSoldier;
-    private Long subdivisionId;
+    private Integer subdivisionId;
     private Boolean isCommander;
 }
